+4XP: moving the Plus4XPProgram and SandProgram into the Renderer and SandPendulum modules. Asserting that parsed number lists that are intended to be vec3s are indeed of length 3. Simplifying the interpolation functions in the animation module with other noclip utilities.
